Change injection point of rules in ComponentStyle
We don't want to inject empty CSS in production, as that
can be a rather slow operation. Instead it's now
only done in development mode, and in production it
only starts injecting rules, once CSS is actually generated.

Note that this doesn't improve performance in the synthetic
benchmarks, but it should have an influence in real-life usage.